requirements.txt
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pika==1.1.0
pycparser==2.19
PyJWT==1.7.1
requests==2.22.0
six==1.12.0
tinydb==3.15.0
urllib3==1.25.6
Werkzeug==0.16.0
